requirements.txt
aiogram==2.25.2
requests
openai-whisper
ffmpeg-python
